+++ lib/asgit.rb
@@ -1,7 +1,7 @@
-require 'asgit/shell'
-require 'asgit/url'
-require 'asgit/config'
-require 'asgit/services'
+require_relative 'asgit/shell'
+require_relative 'asgit/url'
+require_relative 'asgit/config'
+require_relative 'asgit/services'
 
 module Asgit
   class << self
@@ -10,15 +10,15 @@ module Asgit
     # @return [Boolean] true if branch is clean
     def working_tree_clean?
       Shell.run "git status" do |output|
-        return output.include? "nothing to commit"
+        return output.empty?
       end
     end
 
     # Get current git branch based on exec directory
     # @return [String] the current checked out branch
     def current_branch
-      Shell.run "git symbolic-ref HEAD" do |output|
-        return output.strip.gsub(/^refs\/heads\//, '')
+      Shell.run "git symbolic-ref HEAD --short" do |output|
+        return output
       end
     end
 
@@ -33,8 +33,8 @@ module Asgit
     # Check if branch is in sync with remote
     # @return [Boolean]
     def remote_up_to_date?
-      status, stdout, stderr = Shell.run "git push --dry-run"
-      return status && stderr.include?( "Everything up-to-date" )
+      status, stdout, stderr = Shell.run "git push --dry-run --porcelain"
+      return status && stdout.match(/#{current_branch}\s+?\[up\sto\sdate\]/)
     end
 
   end