+++ lib/asgit/services.rb
@@ -6,12 +6,15 @@ module Asgit
         base_url: 'https://github.com',
         base_structure: "%{base_url}/%{organization}/%{project}",
         commit_uri: "commit/%{commit}",
-        branch_uri: "tree/%{branch}"
+        branch_uri: "tree/%{branch}",
+        file_uri: "blob/%{branch}/%{file_path}%{line}"
       },
       bitbucket: {
         base_url: 'https://bitbucket.org',
         base_structure: "%{base_url}/%{organization}/%{project}",
         commit_uri: "commits/%{commit}"
+        branch_uri: "XX",
+        file_uri: "XX"
       }
     }