+++ spec/asgit/url_spec.rb
@@ -14,6 +14,12 @@ describe Asgit::Url do
     Asgit::Services.send(:remove_instance_variable, "@_github")
   end
 
+  describe "::project" do
+    it "returns correct url for the project" do
+      expect( Asgit::Url.project ).to eq "https://github.com/wu/tang"
+    end
+  end
+
   describe "::commit" do
     it "returns the correct url for the provided commit" do
       expect( Asgit::Url.commit "woot" ).to eq "https://github.com/wu/tang/commit/woot"
@@ -23,6 +29,28 @@ describe Asgit::Url do
   describe "::branch" do
     it "returns the correct url for the provided branch" do
       expect( Asgit::Url.branch "woot" ).to eq "https://github.com/wu/tang/tree/woot"
+    end
+  end
+
+  describe "::file" do
+    it "returns the correct url for a file" do
+      expect( Asgit::Url.file "lib/tang.rb" ).to eq "https://github.com/wu/tang/blob/master/lib/tang.rb"
+    end
+
+    it "adjust for fives given with a leading slash" do
+      expect( Asgit::Url.file "/lib/tang.rb" ).to eq "https://github.com/wu/tang/blob/master/lib/tang.rb"
+    end
+
+    it "returns the correct url for a file with a passed branch" do
+      expect( Asgit::Url.file "/lib/tang.rb", branch: 'dev' ).to eq "https://github.com/wu/tang/blob/dev/lib/tang.rb"
+    end
+
+    it "returns the correct url for a file with a line number" do
+      expect( Asgit::Url.file "/lib/tang.rb", branch: 'dev', line: "15" ).to eq "https://github.com/wu/tang/blob/dev/lib/tang.rb#L15"
+    end
+
+    it "returns the correct url for a file with a range of line numbers" do
+      expect( Asgit::Url.file "/lib/tang.rb", branch: 'dev', line: (15..18) ).to eq "https://github.com/wu/tang/blob/dev/lib/tang.rb#L15-L18"
     end
   end