+++ changelog.md
@@ -1,3 +1,17 @@
+# 0.0.5
+- use `--short` flag for current branch (no more parsing!)
+- use `--porcelain` flag for remote_up_to_date?
+- use `--porcelain` flag in working_tree_clean?
+
+# 0.0.4
+- add a default branch option for config
+
+# 0.0.3
+- add a `configured?` boolean method to check if completely configured
+
+# 0.0.2
+- add file link methods
+
 # 0.0.1
 - port status helpers from Statistrano
 - add a config & url helpers